+++ spinnaker/resource_project_test.go
@@ -1,0 +1,75 @@
+package spinnaker
+
+import (
+	"fmt"
+	"net/http"
+	"strconv"
+	"testing"
+	"time"
+
+	"github.com/hashicorp/terraform/helper/acctest"
+	"github.com/hashicorp/terraform/helper/resource"
+	"github.com/hashicorp/terraform/terraform"
+)
+
+func TestAccResourceSourceSpinnakerProject_basic(t *testing.T) {
+	resourceName := "spinnaker_project.test"
+	rName := acctest.RandomWithPrefix("tf-acc-test")
+	resource.ParallelTest(t, resource.TestCase{
+		PreCheck:     func() { testAccPreCheck(t) },
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCheckSpinnakerApplicatioDestroy(resourceName),
+		Steps: []resource.TestStep{
+			{
+				Config: testAccSpinnakerProject_basic(rName),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckSpinnakerProjectExists(resourceName),
+					resource.TestCheckResourceAttr(resourceName, "name", rName),
+					resource.TestCheckResourceAttr(resourceName, "email", "acceptance@test.com"),
+					resource.TestCheckResourceAttr(resourceName, "instance_port", strconv.Itoa(defaultInstancePort)),
+				),
+			},
+		},
+	})
+}
+
+func testAccCheckSpinnakerProjectExists(n string) resource.TestCheckFunc {
+	return func(s *terraform.State) error {
+		rs, ok := s.RootModule().Resources[n]
+		if !ok {
+			return fmt.Errorf("Application not found: %s", n)
+		}
+
+		if rs.Primary.ID == "" {
+			return fmt.Errorf("No Application ID is set")
+		}
+		client := testAccProvider.Meta().(gateConfig).client
+		err := resource.Retry(1*time.Minute, func() *resource.RetryError {
+			_, resp, err := client.ProjectControllerApi.GetUsingGET1(client.Context, rs.Primary.ID)
+			if resp != nil {
+				if resp != nil && resp.StatusCode == http.StatusNotFound {
+					return resource.RetryableError(fmt.Errorf("application does not exit"))
+				} else if resp.StatusCode != http.StatusOK {
+					return resource.NonRetryableError(fmt.Errorf("encountered an error getting application, status code: %data-resources", resp.StatusCode))
+				}
+			}
+			if err != nil {
+				return resource.NonRetryableError(err)
+			}
+			return nil
+		})
+		if err != nil {
+			return fmt.Errorf("Unable to find Application after retries: %s", err)
+		}
+		return nil
+	}
+}
+
+func testAccSpinnakerProject_basic(rName string) string {
+	return fmt.Sprintf(`
+resource "spinnaker_project" "my_proj" {
+	name  = %q
+	email = "acceptance@test.com"
+}
+`, rName)
+}